[PATCH] ia64: fix unwinder bug in unw_access_gr()
I've found this by inspecting the code: pt_regs_off returns an offset into
struct pt_regs.  Relative to latest bk, 2.5 has the same bug.

+++ arch/ia64/kernel/unwind.c
@@ -324,7 +324,7 @@ unw_access_gr (struct unw_frame_info *info, int regnum, unsigned long *val, char
 		} else {
 			/* access a scratch register */
 			pt = get_scratch_regs(info);
-			addr = (unsigned long *) (pt + pt_regs_off(regnum));
+			addr = (unsigned long *) ((unsigned long)pt + pt_regs_off(regnum));
 			if (info->pri_unat_loc)
 				nat_addr = info->pri_unat_loc;
 			else
